fix(python-bildungslogin): restrict access on config.ini
The file is read in several services and scripts. All of these are run by root, except the directory
listener. That service drops privileges and runs as user `listener`. Instead of having read access
by everyone on that file we change the owner to `listener` and permissions to `640`.

I moved away from using dh_fixperms as that would only fix the permissions in the package. On
systems where the package is already installed the permissions would not have been adapted.

issue: ET-38